+++ pipeline/parse/parsers/DiscordParser.ts
@@ -1,9 +1,10 @@
 import { AttachmentType, getAttachmentTypeFromFileName } from "@pipeline/Attachments";
 import { Progress } from "@pipeline/Progress";
-import { ChannelType, RawID } from "@pipeline/Types";
+import { ChannelType } from "@pipeline/Types";
 import { FileInput, streamJSONFromFile } from "@pipeline/parse/File";
 import { JSONStream } from "@pipeline/parse/JSONStream";
 import { Parser } from "@pipeline/parse/Parser";
+import { RawID } from "@pipeline/parse/Types";
 
 export class DiscordParser extends Parser {
     private lastGuildId?: RawID;
