+++ pipeline/process/DatabaseBuilder.ts
@@ -1,10 +1,10 @@
 import { Env } from "@pipeline/Env";
 import { Day, genTimeKeys } from "@pipeline/Time";
-import { Index, RawID, ReportConfig } from "@pipeline/Types";
+import { Index, ReportConfig } from "@pipeline/Types";
 import { createParser } from "@pipeline/parse";
 import { FileInput } from "@pipeline/parse/File";
 import { Parser } from "@pipeline/parse/Parser";
-import { PAuthor, PChannel, PGuild } from "@pipeline/parse/Types";
+import { PAuthor, PChannel, PGuild, RawID } from "@pipeline/parse/Types";
 import { ChannelMessages, ProcessGroupFn } from "@pipeline/process/ChannelMessages";
 import { IndexedMap } from "@pipeline/process/IndexedMap";
 import { MessageProcessor } from "@pipeline/process/MessageProcessor";
